skip rc* replay except rc1

+++ get_diff.sh
@@ -145,8 +145,14 @@ if grep -q "sync_end" result_${start_day}.log && ! grep -q "kill_time" result_${
             echo "Unknown environment: ${env}"
             exit 1
         fi
-        sleep 60
+
         ckb_version=$(sed -n '1p' result_${start_day}.log | grep -oP 'ckb \K[^ ]+(?=\s*\()')
+        if [[ "$ckb_version" == *"rc"* && ! "$ckb_version" =~ rc1$ ]]; then
+            echo "$ckb_version contains 'rc' but does not end with 'rc1'. Exiting..." >>diff_${start_day}.log
+            exit 0
+        fi
+
+        sleep 60
         log_file="block_verifier_${ckb_version}_${env}.log"
         if [ ! -f "$log_file" ]; then
             sudo rm -rf ./replay
